requirements.txt
Django
djangorestframework
psycopg2
django-environ
gunicorn
djangorestframework_simplejwt
requests